+++ lib/knossos.rb
@@ -4,6 +4,7 @@ require 'knossos/display'
 require 'knossos/error'
 require 'knossos/grid'
 require 'knossos/png_adapter'
+require 'knossos/renderer'
 require 'knossos/version'
 
 module Knossos