+++ lib/knossos/renderer.rb
@@ -1,0 +1,6 @@
+require_relative 'renderer/text'
+
+module Knossos
+  class Renderer
+  end
+end